Fix #598 - fix bug where listening mode stopped working
Listening mode would sometimes stop working if the user quickly switched between the Listening category and another category. This was caused by the fact that, when switching back, a new `ListeningResponseViewController` was initialized before the old one was deinitialized. So the `SpeechRecognitionController` was still receiving signals from the "old" insance, which were interfering with the signals from the "new" one.

This was resolved by adding an activate/deactivate mechanism to the `ListeningResponseViewController`, in which it can setup and tear down its relationship with the `SpeechRecognitionController`.